Refactor and fixes
