--- src/components/timeline/duration.ts
@@ -1,8 +1,0 @@
-// From https://github.com/d3/d3-time/blob/main/src/duration.js
-export const durationSecond: number = 1000;
-export const durationMinute: number = durationSecond * 60;
-export const durationHour: number = durationMinute * 60;
-export const durationDay: number = durationHour * 24;
-export const durationWeek: number = durationDay * 7;
-export const durationMonth: number = durationDay * 30;
-export const durationYear: number = durationDay * 365;
