+++ src/utilities/timeline.ts
@@ -1,10 +1,95 @@
+import { bisector, tickStep } from 'd3-array';
 import type { Quadtree, QuadtreeLeaf } from 'd3-quadtree';
 import { scaleLinear, scaleTime, type ScaleLinear, type ScaleTime } from 'd3-scale';
+import {
+  timeHour,
+  timeInterval,
+  timeMillisecond,
+  timeMinute,
+  timeMonth,
+  timeSecond,
+  timeWeek,
+  timeYear,
+  type CountableTimeInterval,
+  type TimeInterval,
+} from 'd3-time';
 
 export enum TimelineLockStatus {
   Locked = 'Locked',
   Unlocked = 'Unlocked',
   TemporaryUnlock = 'TemporaryUnlock',
+}
+
+// From https://github.com/d3/d3-time/blob/main/src/duration.js
+export const durationSecond: number = 1000;
+export const durationMinute: number = durationSecond * 60;
+export const durationHour: number = durationMinute * 60;
+export const durationDay: number = durationHour * 24;
+export const durationWeek: number = durationDay * 7;
+export const durationMonth: number = durationDay * 30;
+export const durationYear: number = durationDay * 365;
+
+// Use a custom D3 time day to force equidistant time intervals
+// for days as opposed to D3's non-uniform intervals that can end early
+// on months or years
+// See https://github.com/d3/d3-scale/issues/245
+// And https://observablehq.com/d/906f777c9f2f0701
+export const customD3TimeDay: CountableTimeInterval = timeInterval(
+  date => date.setHours(0, 0, 0, 0),
+  (date, step) => date.setDate(date.getDate() + step),
+  (start, end) => (end.getTime() - start.getTime()) / durationDay,
+  date => Math.floor(date.getTime() / durationDay),
+);
+
+// TODO may need custom hour, week, month?
+// From https://github.com/d3/d3-time/blob/main/src/ticks.js
+export const customD3TickIntervals: [CountableTimeInterval, number, number][] = [
+  [timeSecond, 1, durationSecond],
+  [timeSecond, 5, 5 * durationSecond],
+  [timeSecond, 15, 15 * durationSecond],
+  [timeSecond, 30, 30 * durationSecond],
+  [timeMinute, 1, durationMinute],
+  [timeMinute, 5, 5 * durationMinute],
+  [timeMinute, 15, 15 * durationMinute],
+  [timeMinute, 30, 30 * durationMinute],
+  [timeHour, 1, durationHour],
+  [timeHour, 3, 3 * durationHour],
+  [timeHour, 6, 6 * durationHour],
+  [timeHour, 12, 12 * durationHour],
+  [customD3TimeDay, 1, durationDay],
+  [customD3TimeDay, 2, 2 * durationDay],
+  [timeWeek, 1, durationWeek],
+  [timeMonth, 1, durationMonth],
+  [timeMonth, 3, 3 * durationMonth],
+  [timeYear, 1, durationYear],
+];
+
+// Based on https://github.com/d3/d3-time/blob/main/src/ticks.js
+export function customD3TickInterval(start: Date, stop: Date, count: number): TimeInterval {
+  // Note: Coerce dates to numbers for arithmetic to make TS happy
+  const target: number = Math.abs(+stop - +start) / count;
+  const i = bisector(([, , step]) => step).right(customD3TickIntervals, target);
+  if (i === customD3TickIntervals.length) {
+    return timeYear.every(tickStep(+start / durationYear, +stop / durationYear, count));
+  }
+  if (i === 0) {
+    return timeMillisecond.every(Math.max(tickStep(+start, +stop, count), 1));
+  }
+  const [t, step] =
+    customD3TickIntervals[target / customD3TickIntervals[i - 1][2] < customD3TickIntervals[i][2] / target ? i - 1 : i];
+  return t.every(step);
+}
+
+// Based on https://github.com/d3/d3-time/blob/main/src/ticks.js
+export function customD3Ticks(start: Date, stop: Date, count: number) {
+  const reverse = stop < start;
+  if (reverse) {
+    [start, stop] = [stop, start];
+  }
+  const interval = customD3TickInterval(start, stop, count);
+  // Make end date inclusive by creating a new date +1ms from stop date
+  const ticks = interval ? interval.range(start, new Date(+stop + 1)) : []; // inclusive stop
+  return reverse ? ticks.reverse() : ticks;
 }
 
 export const CANVAS_PADDING_X = 0;
